adding specification recipe and user logic

+++ javaBackend/src/main/java/com/esewa/javabackend/module/UserStats.java
@@ -3,6 +3,8 @@ package com.esewa.javabackend.module;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.util.UUID;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -14,7 +16,7 @@ public class UserStats {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // separate PK for stats table
+    private Integer id;
 
     @OneToOne
     @JoinColumn(name = "user_id", nullable = false, unique = true)
